Implement unit test for concurrent access to disk.
The disk manager now can be accessed by multiple threads safely.
Different threads reading from the same page on disk with .read_page()
is safe, and different threads writing to different pages with
.write_page() is safe. The disk manager doesn't perform any built-in
locking, so different threads writing to the same file simultaneously
would evidently cause problems.

This isn't an issue in practice, because any request to the disk
manager is handled by the buffer manager, which only grants write
permission to threads which acquire a latch on the specified page.